step 14, exercise 5: ask user to inpuut filenames for loading and saving

+++ directory.rb
@@ -38,24 +38,38 @@ def try_load_students
 end
 
 def load_students(filename = "students.csv")
-  file = File.open(filename, "r")
-  file.readlines.each do |line|
-  name, cohort = line.chomp.split(",")
-    @students << {name: name, cohort: cohort.to_sym}
+  puts "Please enter the file you want to load from"
+  filename = STDIN.gets.chomp
+  if File.exists?(filename)
+    file = File.open(filename, "r")
+    file.readlines.each do |line|
+      name, cohort = line.chomp.split(",")
+      @students << {name: name, cohort: cohort.to_sym}
+    end
+    file.close
+    puts "#{@students.count} students were loaded"
+  else
+    puts "Sorry, #{filename} doesn't exist"
   end
-  file.close
 end
 
 def save_students
-  # open the file for writing
-  file = File.open("students.csv", "w")
-  # iterate over the array of students
-  @students.each do |student|
-    # put student info in an array then join with commas
-    csv_line = [student[:name], student[:cohort]].join(",")
-    file.puts csv_line
+  puts "Please enter the file you want to save to"
+  filename = STDIN.gets.chomp
+  if File.exists?(filename)
+    # open the file for writing
+    file = File.open(filename, "w")
+    # iterate over the array of students
+    @students.each do |student|
+      # put student info in an array then join with commas
+      csv_line = [student[:name], student[:cohort]].join(",")
+      file.puts csv_line
+      file.close
+      puts "#{@students.count} students were saved"
+    end
+  else
+    puts "Sorry, #{filename} doesn't exist"
   end
-  file.close
 end
 
 # step 10
@@ -63,8 +77,8 @@ def print_menu
   puts "Please select the option you want from below by entering the number"
   puts "1. Input the students"
   puts "2. Show the students"
-  puts "3. Save the list to students.csv"
-  puts "4. Load the list from students.csv"
+  puts "3. Save the list"
+  puts "4. Load a list"
   puts "9. Exit"
 end
